Reformat sending email flow
